requirements.txt
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.2.4
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
Flask==2.0.1
